feat: all from config moved to client inside package domain
